[Fix] #34  푸시 알림 화면 전환 기능 수정
